+++ src/motor.hh
@@ -9,7 +9,7 @@ pantilthatns::pantilthat *get_pth();
 int check_pth(pantilthatns::pantilthat *pth);
 
 // Move servo to a desired angle
-int move_servo(pantilthatns::pantilthat *pth, size_t servo_id, float angle);
+int move_servo(pantilthatns::pantilthat *pth, size_t servo_id, int angle);
 
 // Apply an angle with PTH according to the relative position of the target
 struct pos apply_angle(pantilthatns::pantilthat *pth, struct pos relative_pos);
